Separate sector parser from Bluesky parser
